extract more basecontent support from theme into workspace
