--- src/ploneintranet/theme/content/configure.zcml
@@ -1,52 +1,0 @@
-<configure
-    xmlns="http://namespaces.zope.org/zope"
-    xmlns:browser="http://namespaces.zope.org/browser"
-    xmlns:i18n="http://namespaces.zope.org/i18n"
-    i18n_domain="ploneintranet.theme">
-
-  <browser:page
-    name="listing_view"
-    for="plone.app.contenttypes.interfaces.IFolder"
-    layer="..interfaces.IIntranetContentLayer"
-    template="templates/document_view.pt"
-    class=".content.ContentView"
-    permission="zope2.View"
-    />
-
-  <browser:page
-    name="document_view"
-    for="plone.app.contenttypes.interfaces.IDocument"
-    layer="..interfaces.IIntranetContentLayer"
-    template="templates/document_view.pt"
-    class=".content.ContentView"
-    permission="zope2.View"
-    />
-
-  <browser:page
-    name="file_view"
-    for="plone.app.contenttypes.interfaces.IFile"
-    layer="..interfaces.IIntranetContentLayer"
-    template="templates/document_view.pt"
-    class=".content.ContentView"
-    permission="zope2.View"
-    />
-
-  <browser:page
-    name="image_view"
-    for="plone.app.contenttypes.interfaces.IImage"
-    layer="..interfaces.IIntranetContentLayer"
-    template="templates/document_view.pt"
-    class=".content.ContentView"
-    permission="zope2.View"
-    />
-
-  <browser:page
-    name="event_view"
-    for="plone.event.interfaces.IEvent"
-    layer="..interfaces.IIntranetContentLayer"
-    template="templates/event_view.pt"
-    class=".content.ContentView"
-    permission="zope2.View"
-    />
-
-</configure>
