+++ src/ploneintranet/workspace/basecontent/configure.zcml
@@ -1,0 +1,72 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:five="http://namespaces.zope.org/five"
+    xmlns:browser="http://namespaces.zope.org/browser"
+    xmlns:i18n="http://namespaces.zope.org/i18n"
+    xmlns:genericsetup="http://namespaces.zope.org/genericsetup"
+    xmlns:plone="http://namespaces.plone.org/plone"
+    i18n_domain="ploneintranet">
+
+  <adapter
+      for="Products.CMFCore.interfaces.IFolderish
+           ploneintranet.theme.interfaces.IPloneIntranetFormLayer
+           plone.dexterity.interfaces.IDexterityFTI"
+      factory=".document.AddView"
+      name="Document"
+      provides="zope.publisher.interfaces.browser.IBrowserPage"
+      />
+
+  <browser:page
+      for="plone.app.contenttypes.interfaces.IDocument"
+      name="edit"
+      class=".document.EditView"
+      permission="cmf.ModifyPortalContent"
+      layer="ploneintranet.theme.interfaces.IThemeSpecific"
+      />
+
+  <browser:page
+    name="listing_view"
+    for="plone.app.contenttypes.interfaces.IFolder"
+    layer="ploneintranet.theme.interfaces.IIntranetContentLayer"
+    template="templates/document_view.pt"
+    class=".baseviews.ContentView"
+    permission="zope2.View"
+    />
+
+  <browser:page
+    name="document_view"
+    for="plone.app.contenttypes.interfaces.IDocument"
+    layer="ploneintranet.theme.interfaces.IIntranetContentLayer"
+    template="templates/document_view.pt"
+    class=".baseviews.ContentView"
+    permission="zope2.View"
+    />
+
+  <browser:page
+    name="file_view"
+    for="plone.app.contenttypes.interfaces.IFile"
+    layer="ploneintranet.theme.interfaces.IIntranetContentLayer"
+    template="templates/document_view.pt"
+    class=".baseviews.ContentView"
+    permission="zope2.View"
+    />
+
+  <browser:page
+    name="image_view"
+    for="plone.app.contenttypes.interfaces.IImage"
+    layer="ploneintranet.theme.interfaces.IIntranetContentLayer"
+    template="templates/document_view.pt"
+    class=".baseviews.ContentView"
+    permission="zope2.View"
+    />
+
+  <browser:page
+    name="event_view"
+    for="plone.event.interfaces.IEvent"
+    layer="ploneintranet.theme.interfaces.IIntranetContentLayer"
+    template="templates/event_view.pt"
+    class=".baseviews.ContentView"
+    permission="zope2.View"
+    />
+
+</configure>
