+++ src/ploneintranet/workspace/basecontent/templates/document_view.pt
