+++ src/ploneintranet/workspace/basecontent/templates/edit_document.pt
