+++ src/ploneintranet/workspace/basecontent/templates/event_view.pt
