--- src/ploneintranet/workspace/browser/basecontent/configure.zcml
@@ -1,27 +1,0 @@
-<configure
-    xmlns="http://namespaces.zope.org/zope"
-    xmlns:five="http://namespaces.zope.org/five"
-    xmlns:browser="http://namespaces.zope.org/browser"
-    xmlns:i18n="http://namespaces.zope.org/i18n"
-    xmlns:genericsetup="http://namespaces.zope.org/genericsetup"
-    xmlns:plone="http://namespaces.plone.org/plone"
-    i18n_domain="ploneintranet">
-
-  <adapter
-      for="Products.CMFCore.interfaces.IFolderish
-           ploneintranet.theme.interfaces.IPloneIntranetFormLayer
-           plone.dexterity.interfaces.IDexterityFTI"
-      factory=".document.AddView"
-      name="Document"
-      provides="zope.publisher.interfaces.browser.IBrowserPage"
-      />
-
-  <browser:page
-      for="plone.app.contenttypes.interfaces.IDocument"
-      name="edit"
-      class=".document.EditView"
-      permission="cmf.ModifyPortalContent"
-      layer="ploneintranet.theme.interfaces.IThemeSpecific"
-      />
-
-</configure>
