+++ src/ploneintranet/workspace/configure.zcml
@@ -17,6 +17,7 @@
   <include package="ploneintranet.activitystream" />
   <include package="ploneintranet.microblog" />
 
+
   <permission
       id="ploneintranet.workspace.ManageWorkspace"
       title="ploneintranet.workspace: Manage workspace">
@@ -25,8 +26,9 @@
     <role name="TeamManager" />
   </permission>
 
-  <include package=".browser" />
+  <include package=".basecontent" />
   <include package=".behaviors" />
+  <include package=".browser" />
 
   <include file="indexers.zcml" />
